Written unit tests for layer forward prop, working on backprop. Stubs to extend MLP to support other losses than MSE

+++ src/MLP.h
@@ -16,7 +16,7 @@ class MLP {
 public:
     using training_pair_t = std::pair< std::vector<T>, std::vector<T> >;
   //desired call syntax :  MLP({64*64,20,4}, {"sigmoid", "linear"},
-  MLP(const std::vector<unsigned int> & layers_nodes,
+  MLP(const std::vector<size_t> & layers_nodes,
       const std::vector<std::string> & layers_activfuncs,
       bool use_constant_weight_init = false,
       T constant_weight_init = 0.5);
@@ -53,7 +53,7 @@ protected:
                      float learning_rate);
 
 private:
-  void CreateMLP(const std::vector<unsigned int> & layers_nodes,
+  void CreateMLP(const std::vector<size_t> & layers_nodes,
                  const std::vector<std::string> & layers_activfuncs,
                  bool use_constant_weight_init,
                  T constant_weight_init = 0.5);
@@ -66,7 +66,7 @@ private:
   size_t m_num_inputs{ 0 };
   int m_num_outputs{ 0 };
   int m_num_hidden_layers{ 0 };
-  std::vector<unsigned int> m_layers_nodes;
+  std::vector<size_t> m_layers_nodes;
   std::vector<Layer<T>> m_layers;
 };
